Refactor out ExtensionTitleBar
